Add files via upload
The third task for PWC was finding why customers are leaving the company and my suggestion based on the analysis was to
increase your phone service capacity to handle fiber optics complaints. This is currently the most common reason why customers are leaving. Additionally, you might consider offering incentives to customers who pay with an electronic check and who have been subscribed for less than a year but are on a month-to-month contract. These incentives could encourage them to stay with your company longer.